limit LTI Registration DELETE to owning account
why
* so that you can't delete Registrations you don't own,
especially ones from Site Admin

refs INTEROP-8999
flag=lti_registrations_page

test plan:
* create an LTI developer key in Site Admin
* find the corresponding LTI Registration
* make an API request to
`/api/v1/accounts/:account_id/lti_registrations/:id`,
substituting your _root_ account ID for account_id, not
Site Admin, and the Registration ID
* it should fail
* make the same API request for site admin or for a registration
local to your root account
* it should succeed

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/365008
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Xander Moffatt <xmoffatt@instructure.com>

+++ app/controllers/lti/registrations_controller.rb
@@ -1390,6 +1390,10 @@ class Lti::RegistrationsController < ApplicationController
   #   curl -X DELETE 'https://<canvas>/api/v1/accounts/<account_id>/lti_registrations/<registration_id>' \
   #        -H "Authorization: Bearer <token>"
   def destroy
+    unless @context == registration.account
+      return render json: { errors: "registration does not belong to account" }, status: :bad_request
+    end
+
     registration.destroy
     render json: lti_registration_json(registration, @current_user, session, @context, includes: %i[account_binding configuration overlay])
   rescue => e
